--- src/main/java/clarity/gay/commands/CommandInfo.java
@@ -1,10 +1,0 @@
-package clarity.gay.commands;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface CommandInfo {
-    String name();
-    String description() default "";
-}
