+++ src/main/java/clarity/gay/commands/api/CommandInfo.java
@@ -1,0 +1,10 @@
+package clarity.gay.commands.api;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CommandInfo {
+    String name();
+    String description() default "";
+}
