+++ src/main/java/clarity/gay/extensions/FileManager.java
@@ -1,8 +1,8 @@
 package clarity.gay.extensions;
 
-import clarity.gay.modules.Module;
-import clarity.gay.modules.ModuleManager;
-import clarity.gay.modules.ui.ClickGui;
+import clarity.gay.modules.api.Module;
+import clarity.gay.modules.impl.ModuleManager;
+import clarity.gay.modules.impl.ui.ClickGui;
 import com.google.gson.*;
 import java.io.*;
 import java.util.*;
@@ -11,6 +11,8 @@ import com.google.common.collect.Sets;
 public class FileManager {
     public static File rootdir = new File("Clarity");
     public static File modules = new File(rootdir, "modules.json");
+    private HashSet<String> blacklist = Sets.newHashSet(ClickGui.class.getName());
+    private List<File> files = new ArrayList<File>();
 
     public void init() {
         if (!rootdir.exists()) {
@@ -45,8 +47,6 @@ public class FileManager {
         }
     }
 
-    private HashSet<String> blacklist = Sets.newHashSet(ClickGui.class.getName());
-
     public boolean isblacklisted(Module m) {
         return blacklist.contains(m.getClass().getName());
     }
@@ -56,15 +56,15 @@ public class FileManager {
     }
     public List<String> getConfigs() {
         List<String> configFiles = new ArrayList<>();
-        if (rootdir.exists() && rootdir.isDirectory()) {
-            File[] files = rootdir.listFiles();
-            if (files != null) {
-                for (File file : files) {
-                    if (file.isFile() && file.getName().endsWith(".json") && !file.getName().equals("modules.json")) {
-                        configFiles.add(file.getName());
-                    }
-                }
-            }
+        if (rootdir.exists() && rootdir.isDirectory() && rootdir.listFiles() != null) {
+            this.files.addAll(List.of(rootdir.listFiles()));
+            files.stream()
+                    .filter( file ->
+                            file.isFile() &&
+                            file.getName().endsWith(".json") &&
+                            !file.getName().equals("modules.json"))
+                    .forEach(file -> configFiles.add(file.getName())
+            );
         }
         return configFiles;
     }
@@ -83,29 +83,27 @@ public class FileManager {
             load.close();
 
             for (Module mod : ModuleManager.modules.values()) {
-                if (!blacklist.contains(mod.getClass().getName())) {
-                    if (mod.isEnabled()) {
-                        mod.setEnabled(false);
-                    }
-                }
+                if (blacklist.contains(mod.getClass().getName())) continue;
+                if (mod.isEnabled())
+                    mod.setEnabled(false);
             }
 
             Iterator<Map.Entry<String, JsonElement>> itr = json.entrySet().iterator();
             while (itr.hasNext()) {
                 Map.Entry<String, JsonElement> entry = itr.next();
                 Module mod = ModuleManager.getModuleByName(entry.getKey());
-                if (mod != null && !blacklist.contains(mod.getClass().getName())) {
-                    JsonObject jsonMod = (JsonObject) entry.getValue();
-                    boolean enabled = jsonMod.get("enabled").getAsBoolean();
-                    if (mod.isEnabled() == enabled) {
-                        continue;
-                    }
-                    int bind = jsonMod.get("bind").getAsInt();
-                    ArrayList<String> modes = new ArrayList<>();
-                    mod.setMode(modes);
-                    mod.setBind(bind);
-                    mod.setEnabled(enabled);
-                }
+                if (mod == null || !blacklist.contains(mod.getClass().getName())) continue;
+
+                JsonObject jsonMod = (JsonObject) entry.getValue();
+                boolean enabled = jsonMod.get("enabled").getAsBoolean();
+                // but u just disabled all enabled modules?
+                if (mod.isEnabled() == enabled) continue;
+
+                int bind = jsonMod.get("bind").getAsInt();
+                ArrayList<String> modes = new ArrayList<>();
+                mod.setModes(modes);
+                mod.setBind(bind);
+                mod.setEnabled(enabled);
             }
         } catch (Exception e) {
             e.printStackTrace();
