--- src/main/java/clarity/gay/modules/Category.java
@@ -1,12 +1,0 @@
-package clarity.gay.modules;
-
-public enum Category {
-    MOVEMENT,
-    COMBAT,
-    UI,
-    WORLD,
-    PLAYER,
-    FUN,
-    RENDER
-
-}
