+++ src/main/java/clarity/gay/modules/api/Category.java
@@ -1,0 +1,12 @@
+package clarity.gay.modules.api;
+
+public enum Category {
+    MOVEMENT,
+    COMBAT,
+    UI,
+    WORLD,
+    PLAYER,
+    FUN,
+    RENDER
+
+}
