--- src/main/java/clarity/gay/modules/combat/Reach.java
@@ -1,4 +1,0 @@
-package clarity.gay.modules.combat;
-
-public class Reach {
-}
