+++ src/main/java/clarity/gay/modules/impl/combat/Reach.java
@@ -1,0 +1,4 @@
+package clarity.gay.modules.impl.combat;
+
+public class Reach {
+}
