+++ src/main/java/clarity/gay/modules/impl/render/ESP.java
@@ -1,0 +1,4 @@
+package clarity.gay.modules.impl.render;
+public class ESP {}
+
+// soon bro soon... (never)
