--- src/main/java/clarity/gay/modules/render/ESP.java
@@ -1,4 +1,0 @@
-package clarity.gay.modules.render;
-public class ESP {}
-
-// soon bro soon... (never)
